'''Function Inputs Class and Purify Online Learners'''
 * Function now stores a DynamicClass "inputsClass" that can be used to create DenseGenericObjects to store function inputs. By default, training and validation data sets are represented as vectors of such "inputsClass" objects.
 * Purify and simplify the interface of OnlineLearner

+++ trunk/projects/Proteins/WorkUnit/BricoBox.h
@@ -42,7 +42,7 @@ protected:
     for (size_t i = 0; i < n; ++i)
       for (size_t j = i + 1; j < n; ++j)
         numBridges += (size_t)bridges->getElement(i, j).getDouble();
-    return numBridges;
+    return (size_t)numBridges;
   }
   
   static bool checkConsistencyOfBridges(SymmetricMatrixPtr bridges)
